Moves settlement, survivor to lib
