+++ web/src/pages/settlement/population/loader.ts
@@ -1,5 +1,5 @@
 import { redirect, type LoaderFunctionArgs } from "react-router";
-import { Get } from "~/lib/request.tsx";
+import { Get } from "~/lib/request.ts";
 
 export async function loadSurvivors({ params }: LoaderFunctionArgs) {
     const res = await Get(`/api/settlements/${params.settlementId}/survivors`);
